+++ test/plugins/auth/auth_base.js
@@ -2,7 +2,7 @@
 const assert = require('node:assert')
 const { describe, it, beforeEach } = require('node:test')
 
-const { Address } = require('address-rfc2821')
+const { Address } = require('../../../address')
 const fixtures = require('haraka-test-fixtures')
 const utils = require('haraka-utils')